Remove duplicated team service
