+++ frontend/src/app/shared/services/team.service.ts
@@ -23,6 +23,17 @@ export class TeamService {
       .pipe(tap((data) => console.log(data)));
   }
 
+  public createTeam(name: string): Observable<Team> {
+    const team = {
+      name: name,
+    } as Team;
+    return this.httpClient.post<any>('api/v1/teams', team);
+  }
+
+  public updateTeam(team: Team): Observable<Team> {
+    return this.httpClient.put<any>('api/v1/teams/' + team.id, team);
+  }
+
   public getQuarter(date = new Date()) {
     let quarterList: OkrQuarter[] = [];
